Support separately getting order infos
Sometimes, the result record number is too many to display for amazon Order Manager feature when you choose a big last days as a search condition,such as 180,365. For this case, it's necessary to search orders with a smaller date scope condition separately. For example, you can search data from 6/1/17 to 6/1/31 first, and then search data from 7/1/17 to 7/31/17, rather than directly search data from 6/1/17 to 7/31/17.  For this purpose, the script made changes to complete such a new feature which can automatically split the lastdays with 30 to batch getting data.